fix: stdin removing Git repository dependency
When using the stdin option as input for a Git commit message there is still a dependency on being in a Git repository and if you are not we will silently error.
Sorting out the logging and who displays what error message is another issue.

https://gitlab.com/DeveloperC/conventional_commits_linter/-/issues/3

Here we are resolving that issue by removing being in a Git repository as a dependency when providing input via stdin.

We have updated the end-to-end tests for input via stdin to switch to an empty temporary directory.
This verifies the functionality of the stdin input option when not operating within a Git repository.